requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
youtube_dl==2021.12.17
PyNaCl==1.5.0
async-timeout==4.0.3
virtualenv==20.16.3
